Split cli methods
