+++ roboswag/__init__.py
@@ -1,8 +1,2 @@
-from importlib.metadata import version
-
+from roboswag.cli import run_roboswag
 from roboswag.model import APIModel
-
-try:
-    __version__ = version("roboswag")
-except Exception:  # pragma: no cover
-    pass
