+++ roboswag/version.py
@@ -1,0 +1,6 @@
+from importlib.metadata import version
+
+try:
+    __version__ = version("roboswag")
+except Exception:  # pragma: no cover
+    pass
